+++ app/src/main/java/com/teamnk/kimiljung/util/BindingAdapter.kt
@@ -4,8 +4,8 @@ import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 
-@BindingAdapter("loadImage")
-fun ImageView.loadImage(profile : String?){
+@BindingAdapter("profileImage")
+fun ImageView.loadImage(profile : String){
     Glide.with(this)
         .load(profile)
         .into(this)
