+++ app/src/main/res/layout/item_comment.xml
@@ -27,7 +27,7 @@
             style="@style/TextViewPostTitle"
             android:layout_marginStart="8dp"
             android:textSize="12sp"
-            android:text="@{commentList.account_id}"
+            android:text="@{commentList.accountId}"
             android:textStyle="normal"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toEndOf="@id/image_item_comment_image_profile"
